Add relation managers for Customers, Orders, Quotes, Invoices, Opportunities, and Tasks
- Implemented InvoicesRelationManager with form and table configurations.
- Created OpportunitiesRelationManager with necessary actions and filters.
- Developed OrdersRelationManager to manage order-related data.
- Added QuotesRelationManager for handling quotes associated with customers.
- Established TasksRelationManager for task management linked to customers.
- Introduced ItemsRelationManager for managing order items.
- Created AuditLogsRelationManager for tracking user actions.
- Implemented BugReportsRelationManager for managing bug reports.
- Developed ChatSessionsRelationManager for handling customer chat sessions.
- Added InteractionsRelationManager for managing customer interactions.
- Established PermissionsRelationManager for managing user permissions.
- Created RolesRelationManager for handling user roles.
- Added migrations for campaign responses and campaigns.
- Implemented OrderSeeder for seeding order data with related items.

+++ database/migrations/2025_10_08_100552_create_order_items_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->foreignIdFor(Order::class)->constrained()->cascadeOnDelete();
             $table->foreignIdFor(Product::class)->nullable()->constrained()->nullOnDelete();
             $table->string('description');
-            $table->decimal('quantity', 10, 2)->default(1);
+            $table->integer('quantity')->default(1);
             $table->decimal('unit_price', 12, 2)->default(0);
             $table->decimal('discount_amount', 12, 2)->default(0);
             $table->decimal('tax_rate', 5, 2)->default(0);
